add book navigation to portfolio page
relates #12
