+++ App.js
@@ -9,7 +9,7 @@ const App = TabNavigator(
   {
     Appointments: { screen: Appointments },
     Portfolio: { screen: Portfolio },
-    Book: { screen: Book },
+    // Book: { screen: Book },
     Stats: { screen: Stats },
   },
   {
